+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Changed
+- Removed `unwrap()` usage in `ds323x` and `clock`, instead returning `Results` that get carried up to main.
+  - Implemented `std::fmt::Display` and `std::error::Error` for `ds323x`.
+
 ## [1.1.5] - 2026-07-04
 
 ## Removed